blake2: silence two more warnings on platforms with size_t < uint64_t. Don't use SSE2 when cross-compiling

+++ Modules/_blake2/impl/blake2s-ref.c
@@ -298,8 +298,8 @@ int blake2s_update( blake2s_state *S, const uint8_t *in, uint64_t inlen )
     }
     else /* inlen <= fill */
     {
-      memcpy( S->buf + left, in, inlen );
-      S->buflen += inlen; /* Be lazy, do not compress */
+      memcpy( S->buf + left, in, (size_t)inlen );
+      S->buflen += (size_t)inlen; /* Be lazy, do not compress */
       in += inlen;
       inlen -= inlen;
     }
